requirements.txt
Django>=4.2
djangorestframework>=3.14

# Database
psycopg2-binary>=2.9

# AI (Gemini)
google-generativeai>=0.5
google-genai


# File processing
PyPDF2>=3.0
pycryptodome
# Environment variables
python-dotenv>=1.0
